backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
aiofiles>=23.0.0
